gausium_open_platform_connector/pyproject.toml
# SPDX-FileCopyrightText: 2025 InOrbit, Inc.
#
# SPDX-License-Identifier: MIT

[project]
name = "inorbit_gausium_connector"
version = "1.0.9"
description = "InOrbit Connector for Gausium robots managed by Gausium Open Platform."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "InOrbit Inc.", email = "support@inorbit.ai" }]
keywords = [
    "connector",
    "edge-sdk",
    "inorbit",
    "robops",
    "inorbit_gausium_connector",
    "Gausium",
    "Phantas",
]
classifiers = [
    "Intended Audience :: Other Audience",
    "Programming Language :: Python :: 3",
    "Natural Language :: English",
]
dependencies = [
    "inorbit-connector~=1.3.1",
    "httpx>=0.28.1",
    "tenacity>=9.1.2",
    "pytz>=2022.7",
    "ruamel.yaml>=0.18,<0.19",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "python-dotenv==1",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "twine==4.0",
    "build==1.0",
    "bump-my-version==0.15",
    "black",
    "flake8",
    "pytest>=3",
    "pytest-asyncio>=1.0.0",
    "pytest-httpx>=0.35.0",
    "pytest-cov>=6",
    "tox>=4.11.3",
]

[project.urls]
Homepage = "https://github.com/inorbit-ai/inorbit-robot-connectors/tree/main/gausium_open_platform_connector"

[project.scripts]
inorbit_gausium_connector = "inorbit_gausium_connector.inorbit_gausium_connector:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["inorbit_gausium_connector"]

[tool.coverage.run]
source = ["inorbit_gausium_connector/"]

[tool.coverage.report]
fail_under = 0

[tool.flake8]
exclude = [".venv*", "venv", ".tox/"]
max-line-length = 100
show-source = true
statistics = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 100
target-version = ["py312", "py313"]

[tool.tox]
env_list = ["py312", "py313"]

[tool.tox.gh]
python = { "3.12" = "py312", "3.13" = "py313" }

[tool.tox.env_run_base]
extras = ["dev"]
allowlist_externals = ["flake8", "black", "pytest", "coverage"]
commands = [
    [
        "flake8",
        "inorbit_gausium_connector",
        "--max-line-length",
        "100",
    ],
    [
        "black",
        "--check",
        "--diff",
        "--line-length",
        "100",
        ".",
    ],
    [
        "pytest",
        "--cov=.",
    ],
]

[tool.bumpversion]
current_version = "1.0.9"
commit = true
tag = true
message = "Bump inorbit_gausium_open_platform_connector version: {current_version} → {new_version}"
tag_name = "inorbit_gausium_open_platform_connector-v{new_version}"
tag_message = "Bump inorbit_gausium_open_platform_connector version: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "inorbit_gausium_connector/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[[tool.bumpversion.files]]
filename = "docker/examples/run.sh"
search = 'CONNECTOR_VERSION="${{CONNECTOR_VERSION:-"{current_version}"}}"'
replace = 'CONNECTOR_VERSION="${{CONNECTOR_VERSION:-"{new_version}"}}"'

[[tool.bumpversion.files]]
filename = "docker/README.md"
search = '{current_version}'
replace = '{new_version}'
